% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Encryption.R
\name{compressAndEncryptFolder}
\alias{compressAndEncryptFolder}
\title{Compress and encrypt a folder}
\usage{
compressAndEncryptFolder(sourceFolder, targetFileName, publicKeyFileName)
}
\arguments{
\item{sourceFolder}{Name of the folder that must be encrypted.}

\item{targetFileName}{Name of the file that will hold the encrypted data.}

\item{publicKeyFileName}{Name of the file where the public key is stored.}
}
\description{
Compress and encrypt a folder
}
\details{
Compresses all files in a folder and its subfolders, and encrypts using the provided public key.
}
\examples{
\dontrun{
generateKeyPair("public.key", "private.key")

# Create a folder with some data
dir.create("test")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "test/data1.rds")
saveRDS(data, "test/data2.rds")

compressAndEncryptFolder("test", "data.zip.enc", "public.key")
decryptAndDecompressFolder("data.zip.enc", "test2", "private.key")
}
}

