% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encryption.R
\name{compressFolder}
\alias{compressFolder}
\title{Compress a folder}
\usage{
compressFolder(sourceFolder, targetFileName)
}
\arguments{
\item{sourceFolder}{Name of the folder that must be compressed.}

\item{targetFileName}{Name of the file that will hold the compressed data.}
}
\description{
Compress a folder
}
\details{
Compresses all files in a folder and its subfolders, and stores it in a single zip file.
}
\examples{
\dontrun{
# Create a folder with some data
dir.create("test")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "test/data1.rds")
saveRDS(data, "test/data2.rds")

compressFolder("test", "data.zip")
decompressFolder("data.zip", "test2")
}

}

