% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{createResultsDataModelTables}
\alias{createResultsDataModelTables}
\title{Creates the results data model tables to use when uploading results}
\usage{
createResultsDataModelTables(
  connection = NULL,
  connectionDetails = NULL,
  schema,
  resultsFolder
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{schema}{The target schema to create the tables in the database}

\item{resultsFolder}{The folder that holds the resultsDataModelSpecification.csv
file.}
}
\description{
This function assumes that you have used \code{unzipResults} to extract
the contents of your results .zip file and that the file 
resultsDataModelSpecification.csv exists in the resultsFolder. This function
will use resultsDataModelSpecification.csv to create the tables in the schema 
specified.
}
