% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encryption.R
\name{decompressFolder}
\alias{decompressFolder}
\title{Decompress a folder}
\usage{
decompressFolder(sourceFileName, targetFolder)
}
\arguments{
\item{sourceFileName}{Name of the file that must be decompressed.}

\item{targetFolder}{Name of the folder that will hold the extracted data.}
}
\description{
Decompress a folder
}
\details{
Extracts all compressed files to a folder.
}
\examples{
\dontrun{
# Create a folder with some data
dir.create("test")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "test/data1.rds")
saveRDS(data, "test/data2.rds")

compressFolder("test", "data.zip")
decompressFolder("data.zip", "test2")
}

}
