% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Encryption.R
\name{decryptAndDecompressFolder}
\alias{decryptAndDecompressFolder}
\title{Decrypt and decompress a folder}
\usage{
decryptAndDecompressFolder(sourceFileName, targetFolder, privateKeyFileName)
}
\arguments{
\item{sourceFileName}{Name of the file that must be decrypted.}

\item{targetFolder}{Name of the folder that will hold the unencrypted data.}

\item{privateKeyFileName}{Name of the file where the private key is stored.}
}
\description{
Decrypt and decompress a folder
}
\details{
Decrypts the data using the provided private key and extracts all files to a folder.
}
\examples{
\dontrun{
generateKeyPair("public.key", "private.key")

# Create a folder with some data
dir.create("test")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "test/data1.rds")
saveRDS(data, "test/data2.rds")

compressAndEncryptFolder("test", "data.zip.enc", "public.key")
decryptAndDecompressFolder("data.zip.enc", "test2", "private.key")
}
}

