% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encryption.R
\name{decryptFile}
\alias{decryptFile}
\title{Decrypt a data file}
\usage{
decryptFile(sourceFileName, targetFileName, privateKeyFileName)
}
\arguments{
\item{sourceFileName}{Name of the file that must be decrypted.}

\item{targetFileName}{Name of the file that will hold the unencrypted data.}

\item{privateKeyFileName}{Name of the file where the private key is stored.}
}
\description{
Decrypt a data file
}
\details{
Decrypts the data using the provided private key.
}
\examples{
\dontrun{
generateKeyPair("public.key", "private.key")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "data.rds")
encryptFile("data.rds", "data.rds.enc", "public.key")
decryptFile("data.rds.enc", "data2.rds", "private.key")
}

}
