% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{deleteAllRowsForDatabaseId}
\alias{deleteAllRowsForDatabaseId}
\title{Delete all rows for database id}
\usage{
deleteAllRowsForDatabaseId(connection, schema, tableName, databaseId)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{schema}{The schema on the postgres server where the results table exists}

\item{tableName}{Database table name}

\item{databaseId}{Results source database identifier}
}
\description{
Delete all rows for database id
}
\details{
Only PostgreSQL servers are supported.
}
