% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{deleteAllRowsForPrimaryKey}
\alias{deleteAllRowsForPrimaryKey}
\title{Delete results rows for primary key values from database server tables}
\usage{
deleteAllRowsForPrimaryKey(connection, schema, tableName, keyValues)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{schema}{The schema on the postgres server where the results table exists}

\item{tableName}{Database table name}

\item{keyValues}{Key values of results rows to be deleted}
}
\description{
Delete results rows for primary key values from database server tables
}
\details{
Only PostgreSQL servers are supported.
}
