% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encryption.R
\name{encryptFile}
\alias{encryptFile}
\title{Encrypt a data file}
\usage{
encryptFile(sourceFileName, targetFileName, publicKeyFileName)
}
\arguments{
\item{sourceFileName}{Name of the file that must be encrypted.}

\item{targetFileName}{Name of the file that will hold the encrypted data.}

\item{publicKeyFileName}{Name of the file where the public key is stored.}
}
\description{
Encrypt a data file
}
\details{
Encrypts the data using the provided public key.
}
\examples{
\dontrun{
generateKeyPair("public.key", "private.key")
data <- data.frame(x = runif(1000), y = 1:1000)
saveRDS(data, "data.rds")
encryptFile("data.rds", "data.rds.enc", "public.key")
}

}

