% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{getEmptyResultsDataModelSpecification}
\alias{getEmptyResultsDataModelSpecification}
\title{Provides a data.frame for creating a results data model specification}
\usage{
getEmptyResultsDataModelSpecification(verbose = FALSE)
}
\arguments{
\item{verbose}{When TRUE, descriptions of each field in the data.frame
are returned}
}
\description{
This function will provide an empty data model specification data.frame
that defines the minimum required fields for use in the \code{uploadResults}
function.
}
