% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiSharing.R
\docType{package}
\name{OhdsiSharing-package}
\alias{OhdsiSharing}
\alias{OhdsiSharing-package}
\title{OhdsiSharing: Package for sharing of the results of the OHDSI tools}
\description{
Package for sharing of the results of the OHDSI tools, with functions for encrypting results and sending results through SFTP to a central site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiSharing}
  \item \url{https://github.com/OHDSI/OhdsiSharing}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiSharing/issues}
}

}
\author{
\strong{Maintainer}: Marc Suchard \email{msuchard@ucla.edu}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Lee Evans
}

}
\keyword{internal}
