% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{preMergeResultsFiles}
\alias{preMergeResultsFiles}
\title{Pre-merge results files}
\usage{
preMergeResultsFiles(dataFolder, tempFolder = tempdir(), specifications)
}
\arguments{
\item{dataFolder}{folder where the exported results zip files are stored. Zip files containing results
from multiple databases may be placed in the same folder.}

\item{tempFolder}{A folder on the local file system where the zip files are extracted to. Will be cleaned
up when the function is finished. Can be used to specify a temp folder on a drive that
has sufficient space if the default system temp space is too limited.}

\item{specifications}{A tibble data frame object with specifications.}
}
\description{
This function combines results from one or more databases into a single file.
The single file may, for example, be used as the input for a shiny app.
Original location of this function: https://github.com/OHDSI/CohortDiagnostics/blob/93e8915b504318300c68a565945dd17807a88567/R/Shiny.R

It also checks whether the results conform to the results data model specifications.
}
