% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpCd}
\alias{sftpCd}
\title{Change the current working director}
\usage{
sftpCd(sftpConnection, remoteFolder)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}}
function.}

\item{remoteFolder}{The folder on the server to change to.}
}
\description{
Change the current working director
}
