% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpConnect}
\alias{sftpConnect}
\title{Connect to the OHDSI SFTP server}
\usage{
sftpConnect(privateKeyFileName, userName)
}
\arguments{
\item{privateKeyFileName}{A character string denoting the path to an RSA private key.}

\item{userName}{A character string containing the user name.}
}
\value{
An SftpConnection object
}
\description{
Connect to the OHDSI SFTP server
}
