% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpDisconnect}
\alias{sftpDisconnect}
\title{Disconnect from the OHDSI SFTP server.}
\usage{
sftpDisconnect(sftpConnection)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}}
function.}
}
\description{
Disconnect from the OHDSI SFTP server.
}
