% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpGetFile}
\alias{sftpGetFile}
\title{Get a file from the SFTP server}
\usage{
sftpGetFile(sftpConnection, remoteFileName, localFileName = remoteFileName)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}} function.}

\item{remoteFileName}{The name the file on the server.}

\item{localFileName}{The name the file should have locally.}
}
\description{
Get a file from the SFTP server
}
