% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpGetFiles}
\alias{sftpGetFiles}
\title{Get one or more files from the SFTP server}
\usage{
sftpGetFiles(
  sftpConnection,
  remoteFileNames,
  localFolder = getwd(),
  localFileNames = file.path(localFolder, remoteFileNames)
)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}}
function.}

\item{remoteFileNames}{The name of the file(s) to get from the server.}

\item{localFolder}{The path of a local folder where all files will be stored. Is ignored
if localFileNames is provided.}

\item{localFileNames}{The name the file(s) should have locally. If not provided, the files will
be given the same names as on the server.}
}
\description{
Get one or more files from the SFTP server
}
