% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpMkdir}
\alias{sftpMkdir}
\title{Make a directory}
\usage{
sftpMkdir(sftpConnection, remoteFolder)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}} function.}

\item{remoteFolder}{The folder on the server to create.}
}
\description{
Make a directory
}
