% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpPutFile}
\alias{sftpPutFile}
\title{Put a file on the SFTP server}
\usage{
sftpPutFile(
  sftpConnection,
  localFileName,
  remoteFileName = basename(localFileName)
)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}} function.}

\item{localFileName}{The path to the local file to upload.}

\item{remoteFileName}{The name the file should have on the server.}
}
\description{
Put a file on the SFTP server
}
