% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpRename}
\alias{sftpRename}
\title{Rename a file or folder}
\usage{
sftpRename(sftpConnection, oldRemoteFilename, newRemoteFilename)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}} function.}

\item{oldRemoteFilename}{The file on the server to rename.}

\item{newRemoteFilename}{The new file name.}
}
\description{
Rename a file or folder
}
