% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpRm}
\alias{sftpRm}
\title{Remove a file}
\usage{
sftpRm(sftpConnection, remoteFile)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}} function.}

\item{remoteFile}{The file on the server to remove.}
}
\description{
Remove a file
}
