% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpRmdir}
\alias{sftpRmdir}
\title{Remove a directory}
\usage{
sftpRmdir(sftpConnection, remoteFolder)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}}
function.}

\item{remoteFolder}{The folder on the server to remove.}
}
\description{
Remove a directory
}
