% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpUploadFile}
\alias{sftpUploadFile}
\title{Upload a single file to the OHDSI SFTP server}
\usage{
sftpUploadFile(privateKeyFileName, userName, fileName)
}
\arguments{
\item{privateKeyFileName}{A character string denoting the path to an RSA private key.}

\item{userName}{A character string containing the user name.}

\item{fileName}{A character string denoting the path to file to upload.}
}
\description{
This function combines calls to the \code{\link{sftpConnect}}, \code{\link{sftpPutFile}},
and \code{\link{sftpDisconnect}} functions.

A random string will be prefixed to the file name to prevent overwriting existing files on the server.
}
