% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftPwd}
\alias{sftPwd}
\title{Get the present working directory}
\usage{
sftPwd(sftpConnection)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}} function.}
}
\value{
A character string representing the current remote folder name.
}
\description{
Get the present working directory
}
