\name{share}
\alias{share}
\title{share}
\usage{
share(results, ...)
}
\arguments{
  \item{results}{an results object generated by one of the
  OHDSI tools}

  \item{...}{further arguments}
}
\description{
\code{share} copies the results to a central site. By
default, it is sent to the OHDSI coordinating center.
}
\details{
Copies the results to a central site. By default, it is
sent to the OHDSI coordinating center.
}
\examples{
\dontrun{
  connectionDetails <- createConnectionDetails(dbms="sql server", server="RNDUSRDHIT07.jnj.com")
  oscarResults <- oscar(connectionDetails, "cdm4_sim", "scratch", "TestDB")
  share(oscarResults)
}
}

