% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{unzipResults}
\alias{unzipResults}
\title{Unzips a results.zip file and enforces standards required by 
\code{uploadResults}}
\usage{
unzipResults(zipFile, resultsFolder)
}
\arguments{
\item{zipFile}{The location of the .zip file that holds the results to upload}

\item{resultsFolder}{The folder to use when unzipping the .zip file. If this folder
does not exist, this function will attempt to create the folder.}
}
\description{
This function will unzip the zipFile to the resultsFolder and assert
that the file resultsDataModelSpecification.csv exists in the resultsFolder
to ensure that it will work with \code{uploadResults}
}
