% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{uploadFileToS3}
\alias{uploadFileToS3}
\title{Create a public-private key pair}
\usage{
uploadFileToS3(uploadFileName, bucketName, keyName)
}
\arguments{
\item{publicKeyFileName}{Name of the file where the public key should be stored.}

\item{privateKeyFileName}{Name of the file where the private key should be stored.}
}
\description{
Create a public-private key pair
}
\details{
Creates an RSA 4096-bit public-private key pair. The public key can be used to encrypt data, and
only with the private key can the data be decrypted.
}
\examples{
\dontrun{
generateKeyPair("public.key", "private.key")
}

}

