% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{uploadResults}
\alias{uploadResults}
\title{An opinionated function for uploading results a database server.}
\usage{
uploadResults(
  connectionDetails = NULL,
  schema,
  zipFileName,
  createTables = TRUE,
  forceOverWriteOfSpecifications = FALSE,
  purgeSiteDataBeforeUploading = TRUE,
  tempFolder = tempdir()
)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{schema}{The schema on the postgres server where the tables have been created.}

\item{zipFileName}{The path to the zip file containing the results to upload.}

\item{createTables}{When TRUE, the tables specified in the resultsDataModelSpecification.csv
will be created in the schema. NOTE: This will NOT check if the tables
already exist which may cause loss of data.}

\item{forceOverWriteOfSpecifications}{If TRUE, specifications of the phenotypes, cohort definitions, and analysis
will be overwritten if they already exist on the database. Only use this if these specifications
have changed since the last upload.}

\item{purgeSiteDataBeforeUploading}{If TRUE, before inserting data for a specific databaseId all the data for
that site will be dropped. This assumes the input zip file contains the full data for that
data site.}

\item{tempFolder}{A folder on the local file system where the zip files are extracted to. Will be cleaned
up when the function is finished. Can be used to specify a temp folder on a drive that
has sufficient space if the default system temp space is too limited.}
}
\description{
This function will upload results that are in a .ZIP file 
to a database server. This function is opinionated since it enforces the
following requirements for the contents of the .ZIP file:

1. The results stored in the .ZIP file are in .CSV format
2. The name of the result file is the name of the target table in the database
and that the file name is in snake_case.
3. The results .CSV file has column headings in snake_case
4. The column "database_id" is used to indicate the database contributing
the results to the result set.
5. The .ZIP file contains a file called "database.csv" that describes the
database details for the results.
6. The .ZIP file contains a file called "resultsDataModelSpecification.csv" that
fully specifies the results data model.

NOTE: Set the POSTGRES_PATH environmental variable to the path to the folder 
containing the psql executable to enable bulk upload (recommended).
}
