% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{putS3File}
\alias{putS3File}
\title{Put a local file into a remote S3 bucket}
\usage{
putS3File(file, bucket, region = "us-east-1", key, secret,
  appendUuid = TRUE)
}
\arguments{
\item{file}{The path to the file in the local filesystem.}

\item{bucket}{The name of the bucket to put the file in.}

\item{region}{The region of the S3.}

\item{key}{Your AWS access key.}

\item{secret}{Your AWS secret access key.}

\item{appendUuid}{Should append an universally unique identifier to file name when uploading?}
}
\description{
Put a local file into a remote S3 bucket
}

