% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sftp.R
\name{sftpRm}
\alias{sftpRm}
\title{Remove one or more files}
\usage{
sftpRm(sftpConnection, remoteFiles)
}
\arguments{
\item{sftpConnection}{An SftpConnection object as created by the \code{\link{sftpConnect}}
function.}

\item{remoteFiles}{The file(s) on the server to remove.}
}
\description{
Remove one or more files
}
