% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultsDataModel.R
\name{uploadResults}
\alias{uploadResults}
\title{An opinionated function for uploading results a database server.}
\usage{
uploadResults(
  connection = NULL,
  connectionDetails = NULL,
  schema,
  resultsFolder,
  forceOverWriteOfSpecifications = FALSE,
  purgeSiteDataBeforeUploading = TRUE,
  tablePrefix = ""
)
}
\arguments{
\item{connection}{An object of type \code{connection} as created using the
\code{\link[DatabaseConnector]{connect}} function in the
DatabaseConnector package. Can be left NULL if \code{connectionDetails}
is provided, in which case a new connection will be opened at the start
of the function, and closed when the function finishes.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package. Can be left NULL if \code{connection} is
provided.}

\item{schema}{The schema on the postgres server where the tables have been created.
See \code{createResultsDataModelTables} for more information.}

\item{resultsFolder}{The path to the folder containing the results to upload.
See \code{unzipResults} for more information.}

\item{forceOverWriteOfSpecifications}{If TRUE, specifications of the phenotypes, cohort definitions, and analysis
will be overwritten if they already exist on the database. Only use this if these specifications
have changed since the last upload.}

\item{purgeSiteDataBeforeUploading}{If TRUE, before inserting data for a specific databaseId all the data for
that site will be dropped. This assumes the input zip file contains the full data for that
data site.}

\item{tablePrefix}{The table prefix used in naming the tables found in 
resultsDataModelSpecification.csv.}
}
\description{
This function will upload results that were extracted using \code{unzipResults}
and upload them to a database server. This function assumes that database
tables are already created using the \code{createResultsDataModelTables} function.
This function is opinionated since it enforces the following requirements:

1. The results files are in .CSV format
2. The name of the result file is the name of the target table in the database
and that the file name is in snake_case. NOTE: The target table is found in 
"resultsDataModelSpecification.csv" and may include a tablePrefix that is 
specified when uploading results.
3. The results .CSV file has column headings in snake_case
4. The column "database_id" is used to indicate the database contributing
the results to the result set.
6. The results .zip file contains a file called 
"resultsDataModelSpecification.csv" that fully specifies the 
results data model.
}
